File downloading refactor

+++ src/main/java/fun/wilddev/images/services/downloaders/FileDownloader.java
@@ -1,11 +1,8 @@
 package fun.wilddev.images.services.downloaders;
 
-import fun.wilddev.images.exceptions.files.FileDownloadingException;
 import fun.wilddev.images.models.FileMeta;
 import fun.wilddev.images.services.downloaders.callbacks.FileReadyCallback;
 import fun.wilddev.spring.core.services.MessageService;
-
-import java.net.URLConnection;
 
 import lombok.AllArgsConstructor;
 
@@ -13,6 +10,9 @@ import org.springframework.lang.NonNull;
 import org.springframework.util.StreamUtils;
 
 import java.io.*;
+import java.net.*;
+
+import fun.wilddev.images.exceptions.files.*;
 import fun.wilddev.images.services.*;
 
 @AllArgsConstructor
@@ -29,9 +29,12 @@ public class FileDownloader {
     public void download(@NonNull URLConnection conn, @NonNull String id) throws FileDownloadingException {
 
         final String contentType = conn.getContentType();
+        final URL url = conn.getURL();
+
+        File tempFile = null;
 
         try {
-            File tempFile = fileService.createTempFile();
+            tempFile = fileService.createTempFile();
 
             try (InputStream is = new BufferedInputStream(conn.getInputStream());
                  OutputStream os = new BufferedOutputStream(new FileOutputStream(tempFile))) {
@@ -40,13 +43,15 @@ public class FileDownloader {
 
                 gridFsService.store(id, contentType, tempFile);
                 fileReadyCallback.onSuccess(tempFile, new FileMeta(id, contentType));
-
-            } finally {
-                fileService.delete(tempFile);
             }
 
         } catch (Exception ex) {
-            throw new FileDownloadingException(messageService.getMessage("exception.file.downloading.failed", id), ex);
+            throw new FileDownloadingException(messageService
+                    .getMessage("exception.file.downloading.failed", id, url), ex);
+        } finally {
+
+            if (tempFile != null)
+                fileService.delete(tempFile);
         }
     }
 }
